refactor(waveform-seek): H3.4 — extract SeekbarPreview to WaveformSeekPreview.tsx
